Move find_channel_events file into its own file
