+++ trace-to-events/src/processing.rs
@@ -1,23 +1,14 @@
 use crate::{
-    parameters::{
-        AdvancedMuonDetectorParameters, DetectorSettings, FixedThresholdDiscriminatorParameters,
-        Mode, Polarity,
-    },
-    pulse_detection::{
-        advanced_muon_detector::{AdvancedMuonAssembler, AdvancedMuonDetector},
-        threshold_detector::{ThresholdDetector, ThresholdDuration},
-        window::{Baseline, FiniteDifferences, SmoothingWindow, WindowFilter},
-        AssembleFilter, EventFilter, Real, SaveToFileFilter,
-    },
+    channels::find_channel_events, parameters::DetectorSettings, pulse_detection::Real
 };
 use rayon::prelude::*;
 use std::path::{Path, PathBuf};
 use supermusr_common::{
     spanned::{SpanWrapper, Spanned},
-    Channel, EventData, FrameNumber, Intensity, Time,
+    Channel, EventData, FrameNumber
 };
 use supermusr_streaming_types::{
-    dat2_digitizer_analog_trace_v2_generated::{ChannelTrace, DigitizerAnalogTraceMessage},
+    dat2_digitizer_analog_trace_v2_generated::DigitizerAnalogTraceMessage,
     dev2_digitizer_event_v2_generated::{
         finish_digitizer_event_list_message_buffer, DigitizerEventListMessage,
         DigitizerEventListMessageArgs,
@@ -27,191 +18,7 @@ use supermusr_streaming_types::{
 };
 use tracing::info;
 
-#[tracing::instrument(skip_all, fields(channel = trace.channel(), num_pulses))]
-fn find_channel_events(
-    metadata: &FrameMetadataV2,
-    trace: &ChannelTrace,
-    sample_time: Real,
-    detector_settings: &DetectorSettings,
-    save_options: Option<&Path>,
-) -> (Vec<Time>, Vec<Intensity>) {
-    let result = match &detector_settings.mode {
-        Mode::FixedThresholdDiscriminator(parameters) => find_fixed_threshold_events(
-            metadata,
-            trace,
-            sample_time,
-            detector_settings.polarity,
-            detector_settings.baseline as Real,
-            parameters,
-            save_options,
-        ),
-        Mode::AdvancedMuonDetector(parameters) => find_advanced_events(
-            metadata,
-            trace,
-            sample_time,
-            detector_settings.polarity,
-            detector_settings.baseline as Real,
-            parameters,
-            save_options,
-        ),
-    };
-    tracing::Span::current().record("num_pulses", result.0.len());
-    result
-}
-
-#[tracing::instrument(skip_all, level = "trace")]
-fn find_fixed_threshold_events(
-    metadata: &FrameMetadataV2,
-    trace: &ChannelTrace,
-    sample_time: Real,
-    polarity: &Polarity,
-    baseline: Real,
-    parameters: &FixedThresholdDiscriminatorParameters,
-    save_path: Option<&Path>,
-) -> (Vec<Time>, Vec<Intensity>) {
-    let sign = match polarity {
-        Polarity::Positive => 1.0,
-        Polarity::Negative => -1.0,
-    };
-    let raw = trace
-        .voltage()
-        .unwrap()
-        .into_iter()
-        .enumerate()
-        .map(|(i, v)| (i as Real * sample_time, sign * (v as Real - baseline)));
-
-    let pulses = raw
-        .clone()
-        .events(ThresholdDetector::new(&ThresholdDuration {
-            threshold: parameters.threshold,
-            duration: parameters.duration,
-            cool_off: parameters.cool_off,
-        }));
-
-    if let Some(save_path) = save_path {
-        raw.clone()
-            .save_to_file(&get_save_file_name(
-                save_path,
-                metadata.frame_number(),
-                trace.channel(),
-                "raw",
-            ))
-            .unwrap();
-
-        pulses
-            .clone()
-            .save_to_file(&get_save_file_name(
-                save_path,
-                metadata.frame_number(),
-                trace.channel(),
-                "pulses",
-            ))
-            .unwrap();
-    }
-
-    let mut time = Vec::<Time>::new();
-    let mut voltage = Vec::<Intensity>::new();
-    for pulse in pulses {
-        time.push(pulse.0 as Time);
-        voltage.push(pulse.1.pulse_height as Intensity);
-    }
-    (time, voltage)
-}
-
-#[tracing::instrument(skip_all, level = "trace")]
-fn find_advanced_events(
-    metadata: &FrameMetadataV2,
-    trace: &ChannelTrace,
-    sample_time: Real,
-    polarity: &Polarity,
-    baseline: Real,
-    parameters: &AdvancedMuonDetectorParameters,
-    save_path: Option<&Path>,
-) -> (Vec<Time>, Vec<Intensity>) {
-    let sign = match polarity {
-        Polarity::Positive => 1.0,
-        Polarity::Negative => -1.0,
-    };
-    let raw = trace
-        .voltage()
-        .unwrap()
-        .into_iter()
-        .enumerate()
-        .map(|(i, v)| (i as Real * sample_time, sign * (v as Real - baseline)));
-
-    let smoothed = raw
-        .clone()
-        .window(Baseline::new(parameters.baseline_length.unwrap_or(0), 0.1))
-        .window(SmoothingWindow::new(
-            parameters.smoothing_window_size.unwrap_or(1),
-        ))
-        .map(|(i, stats)| (i, stats.mean));
-
-    let events = smoothed
-        .clone()
-        .window(FiniteDifferences::<2>::new())
-        .events(AdvancedMuonDetector::new(
-            parameters.muon_onset,
-            parameters.muon_fall,
-            parameters.muon_termination,
-            parameters.duration,
-        ));
-
-    let pulses = events
-        .clone()
-        .assemble(AdvancedMuonAssembler::default())
-        .filter(|pulse| {
-            Option::zip(parameters.min_amplitude, pulse.peak.value)
-                .map(|(min, val)| min <= val)
-                .unwrap_or(true)
-        })
-        .filter(|pulse| {
-            Option::zip(parameters.max_amplitude, pulse.peak.value)
-                .map(|(max, val)| max >= val)
-                .unwrap_or(true)
-        });
-
-    if let Some(save_path) = save_path {
-        raw.clone()
-            .save_to_file(&get_save_file_name(
-                save_path,
-                metadata.frame_number(),
-                trace.channel(),
-                "raw",
-            ))
-            .unwrap();
-
-        smoothed
-            .clone()
-            .save_to_file(&get_save_file_name(
-                save_path,
-                metadata.frame_number(),
-                trace.channel(),
-                "smoothed",
-            ))
-            .unwrap();
-
-        pulses
-            .clone()
-            .save_to_file(&get_save_file_name(
-                save_path,
-                metadata.frame_number(),
-                trace.channel(),
-                "pulses",
-            ))
-            .unwrap();
-    }
-
-    let mut time = Vec::<Time>::new();
-    let mut voltage = Vec::<Intensity>::new();
-    for pulse in pulses {
-        time.push(pulse.steepest_rise.time.unwrap_or_default() as Time);
-        voltage.push(pulse.peak.value.unwrap_or_default() as Intensity);
-    }
-    (time, voltage)
-}
-
-fn get_save_file_name(
+pub(crate) fn get_save_file_name(
     path: &Path,
     frame_number: FrameNumber,
     channel: Channel,
@@ -244,7 +51,7 @@ pub(crate) fn process<'a>(
 
     let sample_time_in_ns: Real = 1_000_000_000.0 / trace.sample_rate() as Real;
 
-    let vec: Vec<(Channel, _)> = trace
+    let vec: Vec<(Channel, (Vec<u32>, Vec<u16>))> = trace
         .channels()
         .unwrap()
         .iter()
@@ -305,12 +112,14 @@ pub(crate) fn process<'a>(
 
 #[cfg(test)]
 mod tests {
+    use crate::{parameters::{AdvancedMuonDetectorParameters, FixedThresholdDiscriminatorParameters}, Mode, Polarity};
+
     use super::*;
     use chrono::Utc;
+    use supermusr_common::Intensity;
     use supermusr_streaming_types::{
         dat2_digitizer_analog_trace_v2_generated::{
-            finish_digitizer_analog_trace_message_buffer, root_as_digitizer_analog_trace_message,
-            ChannelTraceArgs, DigitizerAnalogTraceMessage, DigitizerAnalogTraceMessageArgs,
+            finish_digitizer_analog_trace_message_buffer, root_as_digitizer_analog_trace_message, ChannelTrace, ChannelTraceArgs, DigitizerAnalogTraceMessage, DigitizerAnalogTraceMessageArgs
         },
         dev2_digitizer_event_v2_generated::{
             digitizer_event_list_message_buffer_has_identifier,
